+++ engine/server/engine/enclave_manager/enclave_manager.go
@@ -97,7 +97,7 @@ func (manager *EnclaveManager) CreateEnclave(
 	shouldDestroyEnclave := true
 	defer func() {
 		if shouldDestroyEnclave {
-			_, destroyEnclaveErrs, err := manager.kurtosisBackend.DestroyEnclaves(teardownCtx, getEnclaveByEnclaveIdFilter(enclaveId));
+			_, destroyEnclaveErrs, err := manager.kurtosisBackend.DestroyEnclaves(teardownCtx, getEnclaveByEnclaveIdFilter(enclaveId))
 			manualActionRequiredStrFmt := "ACTION REQUIRED: You'll need to manually destroy the enclave '%v'!!!!!!"
 			if err != nil {
 				logrus.Errorf("Expected to be able to call the kackend and destroy enclave '%v', but an error occurred:\n%v", enclaveId, err)
@@ -142,6 +142,18 @@ func (manager *EnclaveManager) CreateEnclave(
 		}
 	}()
 
+	var apiContainerHostMachineInfo *kurtosis_engine_rpc_api_bindings.EnclaveAPIContainerHostMachineInfo
+	if apiContainer.GetPublicIPAddress() != nil &&
+		apiContainer.GetPublicGRPCPort() != nil &&
+		apiContainer.GetPublicGRPCProxyPort() != nil {
+
+		apiContainerHostMachineInfo = &kurtosis_engine_rpc_api_bindings.EnclaveAPIContainerHostMachineInfo{
+			IpOnHostMachine:            apiContainer.GetPublicIPAddress().String(),
+			GrpcPortOnHostMachine:      uint32(apiContainer.GetPublicGRPCPort().GetNumber()),
+			GrpcProxyPortOnHostMachine: uint32(apiContainer.GetPublicGRPCProxyPort().GetNumber()),
+		}
+	}
+
 	result := &kurtosis_engine_rpc_api_bindings.EnclaveInfo{
 		EnclaveId:          enclaveIdStr,
 		ContainersStatus:   kurtosis_engine_rpc_api_bindings.EnclaveContainersStatus_EnclaveContainersStatus_RUNNING,
@@ -151,11 +163,7 @@ func (manager *EnclaveManager) CreateEnclave(
 			GrpcPortInsideEnclave:      uint32(apiContainerListenGrpcPortNumInsideNetwork),
 			GrpcProxyPortInsideEnclave: uint32(apiContainerListenGrpcProxyPortNumInsideNetwork),
 		},
-		ApiContainerHostMachineInfo: &kurtosis_engine_rpc_api_bindings.EnclaveAPIContainerHostMachineInfo{
-			IpOnHostMachine:            apiContainer.GetPublicIPAddress().String(),
-			GrpcPortOnHostMachine:      uint32(apiContainer.GetPublicGRPCPort().GetNumber()),
-			GrpcProxyPortOnHostMachine: uint32(apiContainer.GetPublicGRPCProxyPort().GetNumber()),
-		},
+		ApiContainerHostMachineInfo: apiContainerHostMachineInfo,
 	}
 
 	// Everything started successfully, so the responsibility of deleting the enclave is now transferred to the caller
@@ -288,7 +296,7 @@ func (manager *EnclaveManager) getEnclaveApiContainerInformation(
 	*kurtosis_engine_rpc_api_bindings.EnclaveAPIContainerHostMachineInfo,
 	error,
 ) {
-	apiContainerByEnclaveIdFilter := getApiContainerByEnclaveIdFilter(enclaveId);
+	apiContainerByEnclaveIdFilter := getApiContainerByEnclaveIdFilter(enclaveId)
 	enclaveApiContainers, err := manager.kurtosisBackend.GetAPIContainers(ctx, apiContainerByEnclaveIdFilter)
 	if err != nil {
 		return 0, nil, nil, stacktrace.Propagate(err, "An error occurred getting the containers for enclave '%v'", enclaveId)
@@ -302,7 +310,9 @@ func (manager *EnclaveManager) getEnclaveApiContainerInformation(
 		return 0, nil, nil, stacktrace.NewError("Expected to be able to find only one API container associated with enclave '%v', instead found '%v'",
 			enclaveId, numOfFoundApiContainers)
 	}
+
 	apiContainer := getFirstApiContainerFromMap(enclaveApiContainers)
+
 	resultApiContainerStatus, err := getApiContainerStatusFromContainerStatus(apiContainer.GetStatus())
 	if err != nil {
 		return 0, nil, nil, stacktrace.Propagate(err, "An error occurred getting the API container status for enclave '%v'", enclaveId)
@@ -314,13 +324,20 @@ func (manager *EnclaveManager) getEnclaveApiContainerInformation(
 	}
 	var resultApiContainerHostMachineInfo *kurtosis_engine_rpc_api_bindings.EnclaveAPIContainerHostMachineInfo
 	if resultApiContainerStatus == kurtosis_engine_rpc_api_bindings.EnclaveAPIContainerStatus_EnclaveAPIContainerStatus_RUNNING {
-		publicGRPCPort := apiContainer.GetPublicGRPCPort()
-		publicGRPCProxyPort := apiContainer.GetPublicGRPCProxyPort()
-		resultApiContainerHostMachineInfo = &kurtosis_engine_rpc_api_bindings.EnclaveAPIContainerHostMachineInfo{
-			IpOnHostMachine:            apiContainer.GetPublicIPAddress().String(),
-			GrpcPortOnHostMachine:      uint32(publicGRPCPort.GetNumber()),
-			GrpcProxyPortOnHostMachine: uint32(publicGRPCProxyPort.GetNumber()),
+
+		var apiContainerHostMachineInfo *kurtosis_engine_rpc_api_bindings.EnclaveAPIContainerHostMachineInfo
+		if apiContainer.GetPublicIPAddress() != nil &&
+			apiContainer.GetPublicGRPCPort() != nil &&
+			apiContainer.GetPublicGRPCProxyPort() != nil {
+
+			apiContainerHostMachineInfo = &kurtosis_engine_rpc_api_bindings.EnclaveAPIContainerHostMachineInfo{
+				IpOnHostMachine:            apiContainer.GetPublicIPAddress().String(),
+				GrpcPortOnHostMachine:      uint32(apiContainer.GetPublicGRPCPort().GetNumber()),
+				GrpcProxyPortOnHostMachine: uint32(apiContainer.GetPublicGRPCProxyPort().GetNumber()),
+			}
 		}
+
+		resultApiContainerHostMachineInfo = apiContainerHostMachineInfo
 	}
 
 	return resultApiContainerStatus, resultApiContainerInfo, resultApiContainerHostMachineInfo, nil
@@ -395,8 +412,8 @@ func (manager *EnclaveManager) getEnclavesWithoutMutex(
 	}
 
 	result := map[enclave.EnclaveID]*kurtosis_engine_rpc_api_bindings.EnclaveInfo{}
-	for enclaveId, enclave := range enclaves {
-		enclaveInfo, err := manager.getEnclaveInfoForEnclave(ctx, enclave)
+	for enclaveId, enclaveObj := range enclaves {
+		enclaveInfo, err := manager.getEnclaveInfoForEnclave(ctx, enclaveObj)
 		if err != nil {
 			return nil, stacktrace.Propagate(err,"An error occurred getting information about enclave '%v'", enclaveId)
 		}